requirements.txt
roboflow>=1.1.0
opencv-python>=4.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
